refactor: VsfShippingProvider inside ShippingForm
